refactor(shellsafe): extract shared command parser to break duplication
Both the exploit executor and the cleanup executor need the same
quote-aware, metachar-rejecting parser. Keeping two copies meant a fix
to one wouldn't automatically apply to the other.

Extract the logic into internal/shellsafe so both callers import from
one place. exploit/shellparse.go now delegates to shellsafe.Parse;
behavior is identical — all existing shellparse tests still pass.

DefaultCleanupExec drops the sh -c invocation in favour of shellsafe.Parse
+ exec.CommandContext. LLM-generated cleanup commands can no longer abuse
shell features (pipelines, redirects, command substitution) even when a
prompt-injected payload reaches the cleanup registry.
