+++ internal/agent/exploit/shellparse.go
@@ -1,107 +1,14 @@
 package exploit
 
-import (
-	"fmt"
-	"strings"
-	"unicode"
-)
+import "github.com/Armur-Ai/Pentest-Swarm-AI/internal/shellsafe"
 
-// parseCommand splits a command string into argv while respecting single
-// and double quotes. It refuses any command containing shell metacharacters
-// (|, >, <, &, ;, backtick, $(, newline) — callers that need a shell must
-// explicitly wrap their command in `sh -c "..."` as a single quoted arg.
+// ParseCommand splits a command string into argv while respecting single and
+// double quotes. It refuses any command containing shell metacharacters (|, >,
+// <, &, ;, backtick, $(, newlines) unless they appear inside quotes — callers
+// that need a real shell must wrap their command in sh -c "..." explicitly.
 //
-// This is a defence-in-depth layer. Scope validation still runs first.
-// ParseCommand is the exported entry point for callers outside this
-// package (e.g. the ConfirmationAgent re-runs reproduction commands
-// and wants the same safety posture).
+// This is a defence-in-depth layer; scope validation still runs first.
+// ParseCommand is exported so the ConfirmationAgent can use the same policy.
 func ParseCommand(cmd string) ([]string, error) { return parseCommand(cmd) }
 
-func parseCommand(cmd string) ([]string, error) {
-	// Reject obvious shell metacharacters unless they're inside quotes.
-	if err := rejectUnsafe(cmd); err != nil {
-		return nil, err
-	}
-
-	var (
-		args []string
-		cur  strings.Builder
-		in   rune // 0, '\'', or '"'
-		esc  bool // previous char was backslash (inside double quotes only)
-	)
-	flush := func() {
-		if cur.Len() > 0 {
-			args = append(args, cur.String())
-			cur.Reset()
-		}
-	}
-
-	for _, r := range cmd {
-		switch {
-		case esc:
-			cur.WriteRune(r)
-			esc = false
-		case in == '"' && r == '\\':
-			esc = true
-		case in == 0 && (r == '\'' || r == '"'):
-			in = r
-		case in != 0 && r == in:
-			in = 0
-		case in == 0 && unicode.IsSpace(r):
-			flush()
-		default:
-			cur.WriteRune(r)
-		}
-	}
-	if in != 0 {
-		return nil, fmt.Errorf("unterminated %c quote", in)
-	}
-	flush()
-	if len(args) == 0 {
-		return nil, fmt.Errorf("empty command")
-	}
-	return args, nil
-}
-
-// rejectUnsafe scans for shell metacharacters outside of quoted regions.
-// It's intentionally strict — false positives are a feature, not a bug,
-// since any user who needs those characters should wrap in sh -c explicitly.
-func rejectUnsafe(cmd string) error {
-	var (
-		in  rune
-		esc bool
-	)
-	for i, r := range cmd {
-		switch {
-		case esc:
-			esc = false
-			continue
-		case in == '"' && r == '\\':
-			esc = true
-			continue
-		case in == 0 && (r == '\'' || r == '"'):
-			in = r
-			continue
-		case in != 0 && r == in:
-			in = 0
-			continue
-		}
-		if in != 0 {
-			continue
-		}
-		switch r {
-		case '|', '>', '<', '&', ';', '`':
-			return fmt.Errorf("disallowed shell metachar %q at position %d", r, i)
-		case '\n', '\r':
-			return fmt.Errorf("newline in command at position %d", i)
-		case '$':
-			// Reject `$(...)` command substitution. A literal $ followed by
-			// a letter or { is allowed (env-var expansion happens inside exec.Cmd,
-			// not via the shell, so it's inert anyway).
-			if i+1 < len(cmd) && cmd[i+1] == '(' {
-				return fmt.Errorf("command substitution $( at position %d", i)
-			}
-		}
-	}
-	return nil
-}
+func parseCommand(cmd string) ([]string, error) { return shellsafe.Parse(cmd) }
