+++ internal/pipeline/cleanup_memory.go
@@ -2,10 +2,12 @@ package pipeline
 
 import (
 	"context"
+	"fmt"
 	"os/exec"
 	"sync"
 	"time"
 
+	"github.com/Armur-Ai/Pentest-Swarm-AI/internal/shellsafe"
 	"github.com/google/uuid"
 )
 
@@ -85,13 +87,17 @@ func (m *MemoryCleanupRegistry) PendingCleanup(ctx context.Context, campaignID u
 	return out, nil
 }
 
-// DefaultCleanupExec runs a cleanup command string via /bin/sh -c.
-// Cleanup commands are authored by our own agents, not users, so shell
-// expansion here is acceptable — but callers that want stricter semantics
-// can pass a custom exec to NewMemoryCleanupRegistry.
+// DefaultCleanupExec runs a cleanup command without a shell. The command is
+// parsed with the same quote-aware, metachar-rejecting parser the exploit
+// executor uses, so prompt-injected cleanup commands cannot abuse shell
+// features even when the LLM outputs a crafted string.
 func DefaultCleanupExec(ctx context.Context, cmdStr string) error {
+	parts, err := shellsafe.Parse(cmdStr)
+	if err != nil {
+		return fmt.Errorf("unsafe cleanup command: %w", err)
+	}
 	ctx, cancel := context.WithTimeout(ctx, 60*time.Second)
 	defer cancel()
-	cmd := exec.CommandContext(ctx, "sh", "-c", cmdStr)
+	cmd := exec.CommandContext(ctx, parts[0], parts[1:]...)
 	return cmd.Run()
 }
