+++ internal/shellsafe/shellsafe.go
@@ -1,0 +1,97 @@
+// Package shellsafe provides a quote-aware command parser that rejects shell
+// metacharacters outside of quoted regions. Both the exploit executor and the
+// cleanup executor use this so the same safety policy applies to every command
+// the tool fires — LLM-generated or user-authored.
+package shellsafe
+
+import (
+	"fmt"
+	"strings"
+	"unicode"
+)
+
+// Parse splits a command string into argv, respecting single and double quotes.
+// It rejects any unquoted shell metacharacters (|, >, <, &, ;, backtick, $(),
+// newlines). Callers that genuinely need a shell must wrap their command in
+// sh -c "..." as a single quoted argument — that makes the intent explicit.
+func Parse(cmd string) ([]string, error) {
+	if err := rejectUnsafe(cmd); err != nil {
+		return nil, err
+	}
+
+	var (
+		args []string
+		cur  strings.Builder
+		in   rune
+		esc  bool
+	)
+	flush := func() {
+		if cur.Len() > 0 {
+			args = append(args, cur.String())
+			cur.Reset()
+		}
+	}
+
+	for _, r := range cmd {
+		switch {
+		case esc:
+			cur.WriteRune(r)
+			esc = false
+		case in == '"' && r == '\\':
+			esc = true
+		case in == 0 && (r == '\'' || r == '"'):
+			in = r
+		case in != 0 && r == in:
+			in = 0
+		case in == 0 && unicode.IsSpace(r):
+			flush()
+		default:
+			cur.WriteRune(r)
+		}
+	}
+	if in != 0 {
+		return nil, fmt.Errorf("unterminated %c quote", in)
+	}
+	flush()
+	if len(args) == 0 {
+		return nil, fmt.Errorf("empty command")
+	}
+	return args, nil
+}
+
+func rejectUnsafe(cmd string) error {
+	var (
+		in  rune
+		esc bool
+	)
+	for i, r := range cmd {
+		switch {
+		case esc:
+			esc = false
+			continue
+		case in == '"' && r == '\\':
+			esc = true
+			continue
+		case in == 0 && (r == '\'' || r == '"'):
+			in = r
+			continue
+		case in != 0 && r == in:
+			in = 0
+			continue
+		}
+		if in != 0 {
+			continue
+		}
+		switch r {
+		case '|', '>', '<', '&', ';', '`':
+			return fmt.Errorf("disallowed shell metachar %q at position %d", r, i)
+		case '\n', '\r':
+			return fmt.Errorf("newline in command at position %d", i)
+		case '$':
+			if i+1 < len(cmd) && cmd[i+1] == '(' {
+				return fmt.Errorf("command substitution $( at position %d", i)
+			}
+		}
+	}
+	return nil
+}
